Divide stt module
