+++ src/handlers/stt/__init__.py
@@ -1,1 +1,21 @@
-from .stt import * 
+from .stt import STTHandler
+from .custom_handler import CustomSRHandler
+from .sphinx_handler import SphinxHandler
+from .witai_handler import WitAIHandler
+from .googlesr_handler import GoogleSRHandler
+from .vosk_handler import VoskHandler
+from .whisper_handler import WhisperHandler 
+from .groqsr_handler import GroqSRHandler
+from .openaisr_handler import OpenAISRHandler
+
+__all__ = [
+    "STTHandler",
+    "CustomSRHandler",
+    "SphinxHandler",
+    "WitAIHandler",
+    "GoogleSRHandler",
+    "VoskHandler",
+    "WhisperHandler",
+    "GroqSRHandler",
+    "OpenAISRHandler"
+]
