+++ src/window.py
@@ -14,7 +14,6 @@ from .ui import apply_css_to_widget
 from .ui.widgets import MultilineEntry, ProfileRow
 from .constants import AVAILABLE_LLMS, AVAILABLE_PROMPTS, PROMPTS, AVAILABLE_TTS, AVAILABLE_STT
 from gi.repository import Gtk, Adw, Pango, Gio, Gdk, GObject, GLib, GdkPixbuf
-from .handlers.stt import AudioRecorder
 
 from .utility import override_prompts
 from .utility.system import get_spawn_command 
@@ -22,6 +21,7 @@ from .utility.pip import install_module
 from .utility.strings import markwon_to_pango, remove_markdown
 from .utility.replacehelper import replace_variables
 from .utility.profile_settings import get_settings_dict, restore_settings_from_dict
+from .utility.audio_recorder import AudioRecorder
 from .ui.screenrecorder import ScreenRecorder
 
 from .extensions import ExtensionLoader
